JVMCBC-1256: Remove maybeSent
This delivers little real benefit, since we almost immediately
send requests into the GRPC client library, which handles
waiting for the connection to be established for us.  E.g.
at the point we set request.markAsSent(), that library may not
be anywhere close to ready to send it yet.  It may not have
established a connection.

Change-Id: I239d33673fbc3c85faf24c7bde3d1c6dff4ac8ff
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/190294
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>