JVMCBC-1283 SearchNumericRange min/max is zero when it should be null
Motivation
----------
Lets a user distinguish between an unbounded endpoint
and an endpoint bounded at zero.

The SearchNumericRange min() and max() accessors return boxed
Doubles, implying the value is nullable, so this was probably
the intent all along.

Prerequisite for migrating the Kotlin SDK to CoreSearchOps,
since Kotlin validates min <= max if both are present.

Modifications
-------------
Change CoreSearchNumericRange's min and max from primitives
to boxed Doubles. Now they default to null instead of 0.0
when not present in the JSON.

To avoid future confusion, annotate min & max as nullable.
Ditto for CoreSearchDateRange start and end.

Change-Id: I59cf20cf46cf4285a918f9bd1eae24f5c8b79f86
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/190381
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>