Gardening: Fine-tune relative expiry threshold
Motivation
----------
Doesn't matter for users, but it's good for the
code to exactly match the server-side expiry logic,
so developers can look at this code and get an
accurate understanding of how the server behaves.

Modifications
-------------
Express LATEST_VALID_EXPIRY_INSTANT using hex notation,
to make it clear it's the max uint32.

Reduce EARLIEST_VALID_EXPIRY_INSTANT from 31 days to
30 days plus 1 second.

Rename RELATIVE_EXPIRY_CUTOFF_SECONDS to
MAXIMUM_RELATIVE_EXPIRY_SECONDS, so it's clear the
value is inclusive.

When encoding a relative duration for the memcached
protocol, don't convert a duration of exactly 30 days
to an epoch second, because it's not necessary.

Change-Id: I8129f9270292e77887811f532aa0f43f34b371f4
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/190870
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>