JCBC-2069 Add an overload of getAndTouch that takes an Instant
Motivation
----------
Some users want to specify the expiry as an Instant,
and don't want to use the workaround of converting the instant
to a Duration themselves.

The workaround of converting an Instant to a Duration
gives unexpected results if the client's system clock is
set incorrectly.

Modifications
-------------
Add overloads that accept `Instant` instead of `Duration`.

Change-Id: Ib1b4c235e33640308431db084d853c7d09092e06
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/191474
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>