JVMCBC-1309 Protostellar: Fix FTS facet names
Motivation
----------
Facets in Protostellar FTS results were coming back with
empty names.

Modifications
-------------
Get the facet names from the facet map keys, instead of the
*FacetResult's `name` field (which is never populated,
and is soon to be removed from the protocol).

Add and overload of CbCollections.transformValues that takes
a BiFunction (first param is key, second is old value).

Change-Id: I677b11cf627460d06256520dd3c40abac73e999a
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/191767
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>