JCBC-2076 Add an overload of touch that takes an Instant
Motivation
----------
We previously added a similar overload for getAndTouch.

Some users want to specify the expiry as an Instant,
and don't want to use the workaround of converting the instant
to a Duration themselves.

The workaround of converting an Instant to a Duration
gives unexpected results if the client's system clock is
set incorrectly.

Modifications
-------------
Add overloads that accept `Instant` instead of `Duration`.

Change-Id: I48cdf1052a1a56a111e9fef1ed9ec9981a384744
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/192746
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>