JVMCBC-1320 Rework WaitUntilReady
Motivation
----------
Improve reliability of WaitUntilReady (hereafter WUR).

Modifications
-------------
When the desired state is "DEGRADED", require a successful
ping from at least one endpoint of the relevant services.
(Previously, requesting "DEGRADED" caused WUR to time out
if the actual state was "ONLINE".)

When the desired state is "ONLINE", require a successful
ping from at every endpoint of the relevant services.
(Previously, WUR could succeed even if a ping failed,
as long as the diagnostic result considered the channel
connected.)

Explicitly retry failed pings.

Instead of using Flux.interval() as a "pacemaker" to drive
retries, use an actual retry operator with exponential backoff.
This ensures a delay between attempts, and fixes a glitch
where flatMapping the pacemaker emissions was causing
redundant concurrent execution of downstream WUR stages.

WaitUntilReadyContext still includes a diagnostic report.
This report is just informational, and does not control anything.

WaitUntilReadyContext now includes info about pending or
unsuccessful ping targets.

Add an overload for `Deadline.of` that takes a scaling factor,
so it's easy to express a deadline that is some percentage
of another duration. Use this to make the ping timeout 90%
of the WUR timeout, so if ping times out there's a chance
for a retry to succeed.

Upgrade to CouchbaseMock 1.5.27, which adds support for
query and views ping endpoints.

Change-Id: I7a033d1d9dcaee8713e70c71e923b98c34c2c198
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/193558
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>