Gardening: Better exception message when HELO fails
Motivation
----------
The HELO command can fail if the client advertises
an invalid combination of features (for example,
CLUSTERMAP_CHANGE_NOTIFICATION without DUPLEX).

Prior to this change, advertising an invalid combination
would result in an index out of bounds exception when
parsing the response, and the server's error message
would not be surfaced.

Modifications
-------------
If HELO fails, throw an exception whose message
includes the error message returned by the server.

Change-Id: Ibdaa03fc97c82f1cf509b932ab36bdfe574734c7
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/194343
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>