JVMCBC-1334 - Query Delay Disconnect If Outstanding Requests
Motivation: connections of in-progress queries were being closed
if a configuration change removed the query node.

Changes: BaseEndPoint.disconnect() checks for outstanding requests.
If there are outstanding requets, it sets a flag to closeWhenDone
the last request.  If there are no outstandingrequests, it closes
the channel as usual.  Upon completion of request (markRequestCompletion)
if the endpoint is flagged to closeWhenDone and there are no
outstanding requests, the channel is closed.

Change-Id: I93b51b4ed39d790e309ee393004cc7d9f4001b99
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/195619
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>