JVMCBC-1381: Protostellar failing with java.lang.NoClassDefFoundError: io/opentelemetry/context/ContextKey
Fixing a regression introduced in JVMCBC-1314 that is
causing apps without OpenTelemetry jars on the classpath
to fail at runtime.

The goal is to support OpenTelemetry's integration with GRPC so that
Protostellar operations get automatically traced.  This has proven
tricky due to our shading.

The core problem is that we need to connect a user's unshaded
OpenTelemetry object with a shaded GRPC ManagedChannelBuilder.

The io.opentelemetry.instrumentation.opentelemetry-grpc-1.6 library
is used for this, but we need to carefully shade specific portions
of it to account for the unshaded/shaded issue above.

This fix:

* Moves the dependency on opentelemetry-grpc-1.6 out of
  protostellar/pom.xml and into tracing-opentelemetry, where it belongs.
* Except it's not quite that simple.  Since we need to shade this
  library, we need a setup like core-io-deps.  So a new module
  tracing-opentelemetry-deps has to be added.

Change-Id: Id5e316c381be606b73e81f76a75e2dd06acdf604
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/198213
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>