JCBC-2098: BucketSettings will set fields to default values JVMCBC-1395: Protostellar: Cannot create ephemeral or memcached buckets with bucket manager
The Java SDK's BucketSettings sets numerous fields to default values.

BucketSettings is used for both create and update, and this
behaviour seems an outright bug for update, as fields unrelated
to the user's request can get modified.

It likely also causes issues for ephemeral and memcached bucket creation
on Protostellar - but ING-591 complicates that picture.

Ideally we would only be sending settings that the user has
provided.  And if any are mandatory on the server, ideally
those would have to be specified by the user through the SDK.

But that is not where we are, and we cannot change the behaviour
of bucket creation without potentially breaking user's applications.

So the compromise solution is this:

* For bucket creation on classic, things will work exactly as they
  do today.
* For bucket update on classic, we will treat the existing behaviour
  as badly bugged.  Now, only the fields the user is changing, will be
  sent.
  There is potentially risk of breakage here.
* For bucket creation or update on Protostellar, we will use the chance
  for a fresh start, and implement things the way we would prefer them
  to be: we will only send fields that the user has explicitly
  specified.

Change-Id: I095433666124372d48bda754b8c6cb880f6b0300
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/198356
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>