JVMCBC-1393: Protostellar: waitUntilReady() can hang indefinitely
We implement this for Protostellar currently by requesting
a GRPC connection be made with `channel.getState(true)`, and then
using `channel.notifyWhenStateChanged()` to determine when the
channel goes ACTIVE.

Changes in GRPC-Java 1.56.0 mean that if a connection cannot be
established, the channel will no longer cycle through
IDLE -> CONNECTING -> TRANSIENT_FAILURE -> IDLE -> CONNECTING...
Instead it stays in TRANSIENT_FAILURE.  This prevents us
raising a timeout, and the waitUntilReady() call hangs
until the idle timeout (30 minutes by default).

So, a new approach is needed.  We now simply poll
`channel.getState()` until timeout or ACTIVE.
This has been tested with latest GRPC-Java (1.58.0).

Change-Id: Ia00c24509ade4132741bcf5bcc92390af6a09330
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/198480
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>