JCBC-2109: Provide user-configurable control over Timer
Benchmarking tests show that in some high-throughput
scenarios, a single HashedWheelTimer with default config
cannot keep up with the volume of timeouts required.  The
result is that TimeoutExceptions are created much later
than expected.

Along with a series of optimisations occurring in other
reviews, we can also give the user more control over
the internal Timers and HashedWheelTimers to allow them
to optimise for various scenarios.

This change also allows parallel HashedWheelTimers.

The defaults are left unchanged so there is no impact
on existing applications.

Change-Id: Ibb33674cfd521b71f948e62f794954990a4105d3
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/200733
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>