JVMCBC-1426 RangeScanCreate / KeyValueMessageHandler does not retry server busy
Motivation
----------
When the server cannot service a RangeScanCreate request,
it responds with a SERVER_BUSY error code.
Currently, this causes the range scan to fail with a
CouchbaseException. Instead, the client should
retry the request.

Modifications
-------------
When it comes to deciding whether to retry,
treat SERVER_BUSY the same as TEMPORARY_FAILURE.

Change-Id: I32b839a3c3d8b2033fc490965fe16bf70b3c4c2e
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/200892
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>