JVMCBC-1428 - Improvements for Timeouts
Motivation

Expired requests have their cancel() executed synchronously by
the single timer thread and if there are many timeouts, the
cancelling of requests can fall behind resulting in exceptions
being generated later than expected.

Changes

a) AmbiguousTimeoutException and UnambiguousTimeoutException
have stacktraces that are expensive to generate, and not
useful - they just show the timer thread. The improvement
is to construct them without the stacktrace (writableStacktrace=false)

b) request.completeExceptionally(exception) creates a
new CompletionException(exception) from the exception if it
is not a CompletionException. This results in the Throwable
constructor being called, which calls cause.toString() which
in turn calls getMessage() which is very expensive for
CouchbaseExceptions that have an error context.
StacklessException has been added to optimise this path.

Change-Id: I77fef26fafc4355bd150091b1deedc688745c6c8
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/200283
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>