Partial revert "JVMCBC-1428 - Improvements for Timeouts"
This partially reverts commit c6f7fb45509e6f3cf42e589585c00fe14c7d7570.

The CircuitBreaker tests are failing and it appears to be
related to the new StacklessCompletionException.

Initial triage shows that with the change, the circuitBreakerCallback
is being passed a StacklessCompletionException inside BaseEndpoint.send()

While without the change, the original TimeoutException is raised.

The latter causes the circuit breaker to open - the former does not.

Change-Id: I4c454bc930e7e88cd5c14ebf76e20c365b6b3b62
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/201325
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>