Java FIT performer: handle CommandBatch with DEFAULT API using Threads
Currently reactor is used to handle CommandBatch for both
DEFAULT (blocking) and ASYNC (reactive) APIs.

However, reactor isn't well suited for blocking API.  It
means that blocking operations are being done inside a
reactor chain, and it results in wrapped InterruptedExceptions
appearing.

So changing it out for a simple Threading approach.

Also making Hooks.onErrorDropped not set globalError.
In a reactive chain, if one operation fails, all the other
operations are cancelled.  If they subsequently raise
an error this has nowhere to go so is raised on the
global error handling.  This is the standard reactor
UX and shouldn't be regarded as a test failure.

Change-Id: I3297b740420f4042c6078b103b86f973980531a9
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/201694
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>