KCBC-141 Deprecate QueryMetadata.signature in favor of signatureBytes
Motivation
----------
QueryMetadata.signature explodes if the signature is not a JSON Object.

Modifications
-------------
Added new `QueryMetadata.signatureBytes` property. The value is
a byte array containing the signature encoded as JSON.

Deprecate `signature` in favor of `signatureBytes`.

For consistency, add a new `profileBytes` variant of `profile`.
Not deprecating `profile` yet, since we haven't seen problems
with it.

Change-Id: I1b3b5d4f071ee090482c7964b2929e894a30cecd
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/202115
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>