JCBC-2113 Support accessing QueryMetadata.signature when it's not a JSON Object
Motivation
----------
QueryMetadata.signature() explodes if the signature is not a JSON Object.

Modifications
-------------
Added new `QueryMetadata.signatureBytes()` method. The return value is
an optional byte array containing the signature encoded as JSON.

For consistency, add a new `profileBytes()` variant of `profile()`.
Not deprecating `profile()` yet, since we haven't seen problems
with it.

Change-Id: I98b699e116cf7a6c18db2c75e503593529cfe598
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/202173
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: David Nault <david.nault@couchbase.com>