SCBC-441: Transactional queries are NOT_BOUNDED by default
core-io now sends ScanConsistency including when in
NOT_BOUNDED mode, which has revealed bugs in both Java
and now Scala where the default transactional
query ScanConsistency is NOT_BOUNDED, rather than
not-set as it should be.

In general, if the user has not set a query parameter,
we shouldn't be sending it.  So also doing this for
query profile.  This is be a no-op change as
QueryProfile.OFF is the server default currently.

Change-Id: Ic10eb2269ad87e01010c7655052c99b70cb4632b
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/202266
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>