Ensures that parent child relationship is maintained
without this change we're assuming that if there's a span in ThreadLocal and a parent span is being passed to the observation the TracingObservationHandler will assume that the ThreadLocal wins (that's because the user might have created a span manually and we should take that into account).

with this change within the ObservationRequestTracer, we are making the parent span as current before we create its child. That way the parent child relationship is maintained

related to https://issues.couchbase.com/browse/JCBC-2046

Change-Id: I0d21fea31e63ae944d2d814c4b62714889268bf9
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/202961
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>