JVMCBC-1463: Fix tracing-opentelemetry-deps dependency
JVMCBC-1452 made tracing-opentelemetry-deps pull in its transitive
dependencies, to fix runtime failures encountered when trying to use
couchbase2 and OpenTelemetry together.

However, this caused unshaded OpenTelemetry classes to be included
in the tracing-opentelemetry jar.

After further investigation, the cleanest solution is  to
introduce a normal unshaded dependency on the required libraries,
from tracing-opentelemetry.

tracing-opentelemetry already has such a dependency on
io.opentelemetry:opentelemetry-api, so this additional dependency
is not a huge change.

tracing-opentelemetry-deps is now used just for the requirement
of shading specific classes so that we can interface between the
shaded ManagedChannelBuilder that the GRPC library creates, with the
unshaded OpenTelemetry object that the user supplies.

Change-Id: Id76482879a7bf6eb983c41dd1800c4d3509e65e4
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/203028
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>