JVMCBC-1463: Fix tracing-opentelemetry-deps dependency
As an earlier part of this fix, have added the dependencies
of io.opentelemetry.instrumentation:opentelemetry-grpc-1.6 to
the tracing-opentelemetry module.  However, an error is still
being hit:

NoClassDefFoundError: io/opentelemetry/instrumentation/api/instrumenter/rpc/RpcAttributesGetter

This class is in this jar, so adding it to the dependencies:

io.opentelemetry.instrumentation:opentelemetry-instrumentation-api-semconv

tracing-opentelemetry now includes all 3 of
opentelemetry-grpc-1.6's dependencies.

Change-Id: I389a5e7d05b7d9127add25f80fef2229d25e6e6d
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/203135
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>