Add IntMap, an efficient lookup table for integers.
Motivation
----------
This is groundwork for polishing ErrorMap.

IntMap is optimized for small integers like our status codes.
It stores entries in an array unless this would
waste more than 1024 array elements. For sparser keysets,
it falls back to Netty's `IntObjectHashMap`, which is slower
than an array, but faster than HashMap<Integer,?>.

Change-Id: I544f3086f2416aa5903f9f49ab268b9e41a1f1de
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/206054
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>