JVMCBC-1477 (Further) Improve config loading backoff for bad credentials
Motivation
----------
The previous change for JVMCBC-1477 increased the retry delay
for bad credentials from 1 millisecond to 500 milliseconds.
500ms is probably still too short, as it produces two
entries per second in the server's http_access.log.

Modifications
-------------
Increase the fixed "short delay" from 1ms to 10ms.

Use exponential backoff (with the default .5 jitter factor)
for the "long delay" exceptions like bad credentials.

Implement this by chaining `retryWhen` operators.

Consequently, the BucketOpenRetriedEvent is now published
when the retry is _scheduled_, not when it's executed.
Also, the retry delay is not available to include in the event.
These minor changes are acceptable, since the event's
Javadoc says it exists "for debugging reasons".

Change-Id: Iee603ed6aaab105240fd8982963087bda8fd7cc7
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/207419
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>