JCBC-2143 Support Map/List values for "raw" query option
Motivation
----------
The "raw" parameter's method signature uses `Object`
for the value, so it makes sense to allow Maps and List.

Prior to this change, unrecognized types were
silently ignored.

Modifications
-------------
Use JsonValue.coerce() to convert the "raw" map to
JSON. This method handles Maps and Lists as well as
JsonObject/Array.

As a positive side effect, it throws an exception
if an unsupported type is used.

Change-Id: Id44294db292c050bb0341419cca0b95588ead185
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/208925
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>