Test Gardening: Add `Resources` class for loading test resources
Motivation
----------
Better error reporting when the resource is not found.

Makes it so you don't have to specify the class name
every time you load a resource.

Supports reading a resource as an InputStream
or byte array, for cases where those formats are
easier to consume.

Modifications
-------------
Added a copy of the same class from the FIT driver,
without the `Resources.fromThisClass[Loader]()` methods
that require Java 9.

Change-Id: I6fb180fca9bac3d3168f9202aad850ebcd8189f9
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/209047
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>