JVMCBC-1516 PasswordAuthenticator support for atomic dynamic credentials
Motivation
----------
Using separate suppliers for username and password
means it's possible for the backing credentials
to change after the username supplier is called,
but before the password supplier is called,
leading to inconsistent username/password combo.

Modifications
------------
Add `UsernameAndPassword` for representing
credentials as an atomic unit.

Add a builder factory method that takes
`Supplier<UsernameAndPassword>`.

Add a builder factory method that takes
a static username and password.

Deprecate the no-arg builder factory method
and methods for setting username and password.

Change-Id: I4069dcd38a7b644a1d013ce2a9080d4df2a3f580
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/209300
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>