JVMCBC-1517 Analytics Link Manger errors are parsed incorrectly
Motivation
----------
Couchbase Server 7.1.0 started returning errors formatted
as JSON instead of plaintext. That's great, but the SDK
needs to catch up.

Modifications
-------------
When the error parser is given a full HTTP response body
in JSON format, handle that JSON structure, which
is a bit different from what we get out of the
streaming response parser.

Incidentally, if "reason" is present, include it
in toString() because it might be interesting.

Change-Id: I28bade49b0be0c3e01d14f467a0ed4d5f92c7361
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/209701
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>