Gardening: Make BuilderPropertySetter compatible with Columnar
Motivation
----------
Groundwork for Columnar, which exposes the cluster
options sub-builders only via callbacks, and uses
different accessor names.

Modifications
-------------
Make the BuilderPropertySetter class public so it can
be accessed from Columnar packages.

Make the sub-builder accessor name configurable.
Add support for irregular accessor names like
"ioEnvironment", which would be "ioEnvironmentConfig"
if it followed the regular pattern.

When getting a sub-builder, call the accessor that
takes a Consumer (which the builder is passed to)
instead of the deprecated zero-arg method that
returns the builder.

Add a convenience method for setting a whole map of
properties at once.

Change-Id: I9eea17ab1e6fd9c3738ea2798ac7aa9b9e46b770
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/209795
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>