Gardening: Make PartitionInfo.toString() more concise
Motivation
----------
Make it possible for a human to read the partition map.
Reduce noise when logging a bucket config.

Modifications
-------------
Add TopologyHelper.compressKeyRuns(), which turns a map
like this: {0=A, 1=A, 3=A, 4=B, 5=A}
into this: {0..1=A, 3=A, 4=B, 5=A}

Notice the entries for 0 and 1 are merged into a
single "0..1" entry because they are contiguous and
have the same value.

In practical terms, instead of logging a 1024-element
array, we now log something like:
   {0..511=[0,1], {512..1023=[1,0]}

Change-Id: Iaca11e383e1592eff1d41fdd86b64306d280ba75
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/210307
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>