SCBC-462: Scala incorrectly declaring supporting for some transactions extensions
Java and Scala use the same shared implementation of transactions
in core-io.

This currently means that any extensions that are added to Java get
declared in core-io, and hence Scala automatically declares support
for them too.

This is incorrect: while Scala will have the core implementation, it
may not expose the periphery implementation required in the Scala SDK
itself.

This affects several aspects, including that forward compatibility
checks aren't being processed with the correct set of extensions.

This change separates out this idea of supported extensions, and
passes it in from the Java and Scala layers, in the TransactionsConfig.

We also always make sure to pass in a TransactionsConfig to core-io
now, so we can pass this info.

Change-Id: I20ffca45d894c0c92f87052a01a4113db9cc8557
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/211074
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>