Gardening: Replace Owned/ExternalSupplier with OwnedOrExternal
Motivation
----------
Mark a resource as either "owned" or "external" in a simple
way that doesn't require context to understand.

Clarify that the resources previously wrapped in suppliers
aren't really dynamic; the suppliers all returned the same
value every time.

Modifications
-------------
Add `OwnedOrExternal` which associates a boolean "owned"
flag with a resource. Use it instead of the suppliers.

Deprecate `OwnedSupplier` and `ExternalSupplier`.

Annotate OwnedSupplier as used by Spring Data Couchbase,
so not removable.

Not removing ExternalSupplier since it's not marked
as internal API.

Change-Id: I1618e433061a1a63d6499984a5b4511383513616
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/211387
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>