JVMCBC-1527 (Followup) Fix network auto-detection for external addresses with default ports.
Motivation
----------
As part of the topology parser refactor, the network selection
heuristic now requires ports to match (not just the host).

In order for this to work, the SDK must make default ports
explicit before passing seed nodes to the network selector.

Modifications
-------------
Supply default ports so the NetworkSelector reports
a match when connecting to an external network that uses
the default port.

Change-Id: I362ac19069f799be81892f1f26794f43ff11066c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/211874
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>