JVMCBC-1534 DurabilityInvalidLevel in transactions cleanup
Motivation
----------
Some code in transaction land represents "no durability"
as Optional.of(DurabilityLevel.NONE), and uses this
to create a subdoc mutation request.

The code in MemcacheProtocol.mutationFlexibleExtras()
that encodes the durability level expected
Optional.empty() instead, and was incorrectly
encoding a durability requirement with code `0`,
which the server rejected as invalid.

Modifications
-------------
Modify mutationFlexibleExtras to accpet
Optional.of(DurabilityLevel.NONE) as an alternative to
Optional.empty().

Change-Id: I0140d7afacb328ca94d39ecb5abaedea31dd29be
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/212174
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>