Gardening: Accept {} placeholders in transacations log messages
Motivation
----------
Curly braces {} are the standard SLF4J placeholder.
It would be nice if they worked everywhere.

String.format() explodes if there's a type conversion
mismatch like using %d with a String replacement.
Less of that please.

Modifications
-------------
Convert %s and %d into {}, then use SLF4J's
MessageFormatter to do the replacement.

Even with the overhead of %s/%d replacement,
the performance is in line with String.format().

Change-Id: I016ea2d8538bb93128cd088f9b32b566ffe67d21
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/212245
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>