JVMCBC-1542 Service creation failed to use external network host
Motivation
----------
In the wake of JVMCBC-1527 and JVMCBC-1541, Node.addService()
ended up accidentally using the host from the default network
to create services, regardless of which network was actually selected.

Modifications
-------------
Add NodeIdentifier.hostForNetworkConnections(), which returns
the host from the selected network.

Use this instead of NodeIdentifier.address() in Node.addService().

Caveats
-------
NodeIdentifiers created by the legacy GlobalConfigParser and
BucketConfigParser are not assigned a value for the new
`hostForNetworkConnections` property. Those parsers are no longer
used by production SDK code, and are retained for now just in case
other projects reference them.

Change-Id: I1fb8d0d56700b05ead501b14494407db5d15797c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/212914
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>