JVMCBC-1543 ViewLocator should use NodeIdentifier instead of hostname
Motivation
----------
The ViewLocator was using just a node's host to decide whether
it can service view queries. This was potentially inaccurate,
since multiple nodes can run on the same host.

Apart from the inaccuracy, this strategy no longer works now that
alternate addresses are resolved when the config is parsed.
Now is the time to use a more robust strategy.

Modifications
-------------
Compare NodeIdentifiers instead of just hosts.

Change-Id: Iaaf44604d9fdf43eb28441c3a248fdc4a5daa9c6
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/212915
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>