Gardening: Config provider owns topology parser
Motivation
----------
DNS SRV lookup must happen before TopologyParser is created.
Core currently does the DNS SRV lookup synchronously and then
creates the TopologyParser, but for JCO-3 we want to make the
lookup asynchronous like other bits of bootstrap.

The TopologyParser is more at home in the config provider anyway,
since that was its only user.

This also simplifies the unit test, since less of Core needs
to be mocked.

Modifications
-------------
Make DefaultConfigurationProvider (instead of Core) responsible
for creating the TopologyParser.

Change-Id: I0b8a96a3990f25e9288b2238e451cfce2074b8cc
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/213208
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>