JVMCBC-1573: ExtParallelUnstaging leading to OOM when many concurrent transactions executed
`.parallel(1000)`, when run on an elastic thread pool (as we
are doing), results in 1000 threads being created, no matter
how many operations will actually be run in that chain.

This is of course expensive, and leading to OOM issues if many
concurrent transactions are run.  Testing indicates that 200
parallel transactions are using 3.6GiB of memory, when previously
they would use < 200MiB.  Plus, creating 1000 * 1000 threads.

Replacing this with .flatMap(), which leads to threads only
being created as required.  In testing only a small
number are now created for the 1,000 concurrent ops, and memory
usage is back to < 200MiB.

Also adding a backdoor to control parallelism, for emergency
purposes.

Change-Id: I38fcd1f154abfdf6ed8cdf69b6750bede6f0af60
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/216821
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>