JVMCBC-1560: Add cluster UUID and name to metrics and spans (1/n)
Preparing the groundwork for this change, by adding
CoreResources.

Currently all span-created logic accesses the CoreEnvironment's
RequestTracer.

Since a CoreEnvironment may be shared between multiple Clusters,
we cannot easily give Core-specific information to this
RequestTracer.

So a new RequestTracer is required, that lives outside of
CoreEnvironment, belongs to a Core, and wraps the CoreEnvironment's
RequestTracer.

It would have been possible to add this to Core, since
Protostellar does not require the cluster-label functionality
that is initially required.

But it seems a more general and forward-looking approach to
formalise this concept - of resources associated with a Core
or CoreProtostellar, that are now owned by a CoreEnvironment.
So, CoreResources is born.

Change-Id: Ia0d3765cc115b3d5405c89a7c4665d2b134dd295
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/217012
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>