JVMCBC-1577 ClusterTopology refactoring broke serviceShowsUpInConfig()
Motivation
----------
After the refactoring, only one of NodeInfo.services()
or NodeInfo.sslServices() is populated, depending on
whether TLS was used to connect.

Same for PortInfo.ports() / PortInto.sslPorts().

This could cause the SDK to erroneously report
a service is not available, even though it is
actually available.

Modifications
-------------
Update serviceShowsUpInConfig() to look in both places,
instead of just the non-TLS place.

Take advantage of BucketConfig.serviceEnabled() which
does the check correctly (looks in both places).

Change-Id: I63cf3a8a2c174c95e49c4eade433da9c3a42cfc6
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/218411
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>