JCBC-2170 FIT: ClusterEnvironment.publishOnScheduler()
Motivation
----------
Verify the SDK publishes results on the configured
scheduler.

Modifications
-------------
Configure the FIT ClusterEnvironment to publish on a
custom scheduler.

Add UserSchedulerUtil.withSchedulerCheck(Mono/Flux).
It returns a new Mono/Flux with doOn* callbacks
that verify the signals are published on the custom
scheduler.

Everywhere the performer calls a reactive SDK method,
wrap the result in `withSchedulerCheck()`.

Change-Id: I40a5c053603d34ed8e86743f286c91bf77273010
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/219006
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>