Gardening: Use ClusterTopology in Core
Motivation
----------
Migrate to new ClusterTopology model.

Modifications
-------------
DRY: reconfigureGlobal() and reconfigureBuckets() can now delegate
to the same reconfigureGlobalOrBucket() method. The only difference
is whether the services get created at global or bucket scope
(depending on whether the `bucket` argument is null).

DRY in maybeRemoveNode -- code to look for node is now the same
regardless of whether the topology is "global" or came from a bucket.

Change-Id: I1405d480404711071819c8034a519dca5219e6d8
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/219091
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>