Gardening: Create new OrchestratorProxy for each test
Motivation
----------
Groundwork for migrating to new ClusterTopology model.

Makes it easier to use a real ClusterTopology instance
instead of a mock (coming in a future change).

Important because with the new model it's easier
to create a real ClusterTopology than a mock one.

Modifications
-------------
For each test, create a new OrchestratorProxy initialized
with the desired partition data.

Change-Id: Icf4cd42d644f5fe0a5e8daa9108d45f54d51d66a
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/219213
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>