Gardening: Simplify KeyValueLocator.calculateNodeId()
Motivation
----------
Remove special-case logic for requests that target replicas.

Reinforce the idea that the primary ("active") replica
set member shares the same index space as the replicas.
In other words, 0 is the primary, 1 is the first replica, etc.
just like in the bucket topology JSON.

Modifications
-------------
Push `replica()` from ReplicaGetRequest and friends
up into the KeyValueRequest interface, with a default
implementation that says the request targets the primary.

Change-Id: I84159e72f3218af01e3989854909bf84dc362edb
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/219277
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>