JCBC-2175: Separate MpscArrayQueue creation in certain classes
Motivation
----------
For native image compatibility, instances of MpscArrayQueue must be substituted for MpscAtomicUnpaddedArrayQueue from the jctools library.

The affected classes are DefaultEventBus, OrphanReporter and ThresholdLoggingTracer.

While it’s possible to substitute the constructor of a class, it somehow caused unwanted repercussions for the OrphanReporter and ThresholdLoggingTracer classes. To simplify the substitution, a separate static factory method should be added to produce a Queue<T> object, which can be easily substituted in the quarkus extension.

Changes
-------
- Separated Queue assignment in static factory method in NativeImageHelper
- Added UsedBy(QURAKUS_COUCHBASE) annotation
- Added checkstyle for using org.jctools library

Change-Id: I397ded51911c6f4c346a5e2e2d8105a3b1d4ed4d
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/219572
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>